# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the api documentation of task manager project.",
        "title": "Task Manager API Documentation",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:9090",
    "basePath": "/",
    "paths": {
...
        "/api/v1/teams": {
            "get": {
                "description": "Get all teams of user based on query parameters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teams"
                ],
                "summary": "Get all teams",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003caccess_token\u003e",
                        "description": "Access Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "return teams created by you if createdByMe set to true otherwise false.",
                        "name": "createdByMe",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Number of tasks to return per page (default 10)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset for pagination (default 0)",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search term to filter tasks",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Sort tasks by create time (true for ascending, false for descending)",
                        "name": "sortByCreatedAt",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Teams fetched successfully.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.Team"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.CustomError"
                        }
                    },
                    "401": {
                        "description": "Either refresh token not found or token is expired.",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.CustomError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.CustomError"
                        }
                    }
                }
            },
            "post": {
                "description": "CreateTeam API is made for creating a new team in the task manager application.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teams"
                ],
                "summary": "Create New Team",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003caccess_token\u003e",
                        "description": "Access Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "2024-03-25T22:59:59.000Z",
                        "name": "createdAt",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "example": 954751326021189799,
                        "name": "createdBy",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "example": 954751326021189633,
                        "name": "id",
                        "in": "formData"
                    },
                    {
                        "maxLength": 15,
                        "minLength": 3,
                        "type": "string",
                        "example": "Team Jupiter",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "enum": [
                            "PUBLIC",
                            "PRIVATE"
                        ],
                        "type": "string",
                        "example": "PUBLIC",
                        "name": "privacy",
                        "in": "formData"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "collectionFormat": "multi",
                        "description": "Ids of user who will be added to the team.",
                        "name": "members",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Team created successfully.",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.CustomError"
                        }
                    },
                    "401": {
                        "description": "Either refresh token not found or token is expired.",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.CustomError"
                        }
                    },
                    "500": {
                        "description": "Internal server error.",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.CustomError"
                        }
                    }
                }
            }
        },
        "/api/v1/teams/leave/{TeamID}": {
            "delete": {
                "description": "Removes user from particular team",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teams"
                ],
                "summary": "Leave Team",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003caccess_token\u003e",
                        "description": "Access Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID of team whose members you want.",
                        "name": "TeamID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Team left successfully.",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "you are not a member of that team.",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.CustomError"
                        }
                    },
                    "401": {
                        "description": "Either refresh token not found or token is expired.",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.CustomError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.CustomError"
                        }
                    }
                }
            }
        },
        "/api/v1/teams/members": {
            "post": {
                "description": "Add members to a team based on provided parameters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teams"
                ],
                "summary": "Add members to a team",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Team ID",
                        "name": "TeamID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "Bearer \u003caccess_token\u003e",
                        "description": "Access Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "collectionFormat": "multi",
                        "description": "Array of member IDs to add to the team",
                        "name": "memberIds",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Members added successfully",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.CustomError"
                        }
                    },
                    "401": {
                        "description": "Either refresh token not found or token is expired.",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.CustomError"
                        }
                    },
                    "403": {
                        "description": "Not allowed to add members.",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.CustomError"
                        }
                    },
                    "409": {
                        "description": "Member already exist.",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.CustomError"
                        }
                    },
                    "500": {
                        "description": "Internal server error.",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.CustomError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove members from a team based on provided parameters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teams"
                ],
                "summary": "Remove members from a team",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Team ID",
                        "name": "TeamID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "Bearer \u003caccess_token\u003e",
                        "description": "Access Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "collectionFormat": "multi",
                        "description": "Array of member IDs to add to the team",
                        "name": "memberIds",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Members Removed successfully",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.CustomError"
                        }
                    },
                    "401": {
                        "description": "Either refresh token not found or token is expired.",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.CustomError"
                        }
                    },
                    "403": {
                        "description": "Not allowed to remove members.",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.CustomError"
                        }
                    },
                    "500": {
                        "description": "Internal server error.",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.CustomError"
                        }
                    }
                }
            }
        },
        "/api/v1/teams/{TeamID}/members": {
            "get": {
                "description": "Get all members of team based on query parameters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teams"
                ],
                "summary": "Get all team members",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003caccess_token\u003e",
                        "description": "Access Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID of team whose members you want.",
                        "name": "TeamID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Number of tasks to return per page (default 10)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset for pagination (default 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Team members fetched successfully.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.User"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.CustomError"
                        }
                    },
                    "401": {
                        "description": "Either refresh token not found or token is expired.",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.CustomError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.CustomError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "errorhandling.CustomError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "Bad request"
                },
                "error": {
                    "type": "string",
                    "example": "Corresponding Error Message will Show Here"
                }
            }
        },
        "response.SuccessResponse": {
            "description": "Send success response to client with corresponding message and id(if any).",
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "200 OK"
                },
                "id": {
                    "type": "integer",
                    "example": 974751326021189896
                },
                "message": {
                    "type": "string",
                    "example": "Task Created Successfully."
                }
            }
        },
...
        "response.Team": {
            "description": "Team information with it's id, name, privacy (PUBLIC or PRIVATE), id of user who created it, time when it was created and team members.",
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string",
                    "example": "2024-03-25T22:59:59.000Z"
                },
                "createdBy": {
                    "type": "integer",
                    "example": 954751326021189799
                },
                "id": {
                    "type": "integer",
                    "example": 954751326021189633
                },
                "name": {
                    "type": "string",
                    "example": "Team Jupiter"
                },
                "teamPrivacy": {
                    "type": "string",
                    "example": "PUBLIC"
                }
            }
        },
...
        "response.User": {
            "description": "User information with id, first name, last name, bio, email, password and privacy.",
            "type": "object",
            "properties": {
                "bio": {
                    "type": "string",
                    "example": "Junior Software Engineer at ZURU TECH INDIA."
                },
                "email": {
                    "type": "string",
                    "example": "chiragmakwana@gmail.com"
                },
                "firstName": {
                    "type": "string",
                    "example": "Chirag"
                },
                "id": {
                    "type": "integer",
                    "example": 974751326021189896
                },
                "lastName": {
                    "type": "string",
                    "example": "Makwana"
                },
                "password": {
                    "type": "string",
                    "example": "Chirag123$,omitempty"
                },
                "privacy": {
                    "type": "string",
                    "example": "PUBLIC"
                }
            }
        },
...
    }
}